Implementing basic pricelist parser
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C34146

+++ TOneV2/Code/TOneV2/TOne.WhS.SupplierPriceList.Business/Context/SupplierPriceListExecutionContext.cs
@@ -9,6 +9,6 @@ namespace TOne.WhS.SupplierPriceList.Business
 {
     public class SupplierPriceListExecutionContext : ISupplierPriceListExecutionContext
     {
-        public int InputFileId { get; set; }
+        public long InputFileId { get; set; }
     }
 }
